+++ process/process_darwin.go
@@ -99,28 +99,6 @@ func (p *Process) Exe() (string, error) {
 	return p.ExeWithContext(context.Background())
 }
 
-func (p *Process) ExeWithContext(ctx context.Context) (string, error) {
-	lsof_bin, err := exec.LookPath("lsof")
-	if err != nil {
-		return "", err
-	}
-	out, err := invoke.CommandWithContext(ctx, lsof_bin, "-p", strconv.Itoa(int(p.Pid)), "-Fpfn")
-	if err != nil {
-		return "", fmt.Errorf("bad call to lsof: %s", err)
-	}
-	txtFound := 0
-	lines := strings.Split(string(out), "\n")
-	for i := 1; i < len(lines); i += 2 {
-		if lines[i] == "ftxt" {
-			txtFound++
-			if txtFound == 2 {
-				return lines[i-1][1:], nil
-			}
-		}
-	}
-	return "", fmt.Errorf("missing txt data returned by lsof")
-}
-
 // Cmdline returns the command line arguments of the process as a string with
 // each argument separated by 0x20 ascii character.
 func (p *Process) Cmdline() (string, error) {
